Refactor Test Runner - split specs parsing from runner execution
- Resolve specs path using current script path
- Runner and Specs are separated files
- index.js combines the specs, runner and ava-describe
- Cleanup dependencies
